+++ src/app/auth/signup/components/SignUpForm.tsx
@@ -5,7 +5,7 @@ import { LogIn } from 'lucide-react'
 import { useRouter } from 'next/navigation'
 import { useState, useTransition } from 'react'
 
-import { getAdminRole } from '@/actions/roles'
+import { getAdminRole } from '@/actions/roles.actions'
 import { Button } from '@/components/Button'
 import { MessageError } from '@/components/Error'
 import { authClient } from '@/lib/auth-client'
