+++ src/app/dashboard/management/roles/components/role-name.tsx
@@ -4,7 +4,7 @@ import { useAtomValue, useSetAtom } from 'jotai'
 import { SquarePenIcon } from 'lucide-react'
 import { startTransition, useEffect, useRef, useState } from 'react'
 
-import { editRoleName } from '@/actions/roles'
+import { editRoleName } from '@/actions/roles.actions'
 import { Button } from '@/components/Button'
 import { rolesAtom, selectedRoleAtom } from '@/global/management-roles'
 import { cn } from '@/lib/utils'
