+++ src/app/dashboard/management/roles/components/roles-table.tsx
@@ -4,7 +4,7 @@ import { useAtom, useAtomValue, useSetAtom } from 'jotai'
 import { SearchIcon } from 'lucide-react'
 import { useEffect } from 'react'
 
-import { getRoles } from '@/actions/roles'
+import { getRoles } from '@/actions/roles.actions'
 import { Badge } from '@/components/Badge'
 import { Card, CardContent } from '@/components/Card'
 import { SimpleInput } from '@/components/Inputs'
