+++ src/services/roles.effect.ts
@@ -1,12 +1,6 @@
 import { Effect } from 'effect'
 
 import {
-    DB_STATES,
-    DEFAULT_PERMISSIONS,
-    DEFAULT_ROLES,
-} from '@/constants/client'
-
-import {
     InvalidInputError,
     AlreadyExistsError,
     UnexpectedError,
@@ -14,6 +8,11 @@ import {
 } from '../errors'
 import { PrismaService } from '../prisma/db'
 
+import {
+    DB_STATES,
+    DEFAULT_PERMISSIONS,
+    DEFAULT_ROLES,
+} from '@/constants/client'
 
 export const editRoleNameEffect = (id: string, name: string) =>
     Effect.gen(function* (_) {
